requirements.txt
flask
langchain
langchain-community
langchain-huggingface
langchain-groq
rich
python-dotenv
